refactor: 세션 페이지 useSession 훅 분리
