+++ frontend/src/pages/SessionPage.tsx
@@ -1,260 +1,33 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useMemo } from "react";
 import VideoContainer from "@/components/session/VideoContainer.tsx";
-import { useNavigate, useParams } from "react-router-dom";
+import { useParams } from "react-router-dom";
 import SessionSidebar from "@/components/session/SessionSidebar.tsx";
 import SessionToolbar from "@/components/session/SessionToolbar.tsx";
-import useMediaDevices from "@/hooks/useMediaDevices.ts";
-import useToast from "@/hooks/useToast.ts";
-import usePeerConnection from "@/hooks/usePeerConnection.ts";
 import useSocketStore from "@/stores/useSocketStore";
 import useSessionFormStore from "@/stores/useSessionFormStore";
-
-interface User {
-  id: string;
-  nickname: string;
-}
+import { useSession } from "@/hooks/useSession";
 
 const SessionPage = () => {
-  const { socket, connect } = useSocketStore();
+  const { socket } = useSocketStore();
   const { sessionName } = useSessionFormStore();
-
-  const {
-    createPeerConnection,
-    closePeerConnection,
-    peers,
-    setPeers,
-    peerConnections,
-  } = usePeerConnection(socket!);
   const { sessionId } = useParams();
-  const [nickname, setNickname] = useState<string>("");
-  const [reaction, setReaction] = useState("");
-
   const {
+    nickname,
+    setNickname,
+    reaction,
+    peers,
     userVideoDevices,
     userAudioDevices,
-    selectedAudioDeviceId,
-    selectedVideoDeviceId,
-    stream,
     isVideoOn,
     isMicOn,
+    stream,
     handleMicToggle,
     handleVideoToggle,
     setSelectedAudioDeviceId,
     setSelectedVideoDeviceId,
-    getMedia,
-  } = useMediaDevices();
-
-  const reactionTimeouts = useRef<{
-    [key: string]: ReturnType<typeof setTimeout>;
-  }>({});
-  const navigate = useNavigate();
-  const toast = useToast();
-
-  useEffect(() => {
-    if (!socket) connect(import.meta.env.VITE_SIGNALING_SERVER_URL);
-    const connections = peerConnections;
-
-    return () => {
-      Object.values(connections.current).forEach((pc) => {
-        pc.ontrack = null;
-        pc.onicecandidate = null;
-        pc.oniceconnectionstatechange = null;
-        pc.onconnectionstatechange = null;
-        pc.close();
-      });
-    };
-  }, []); // eslint-disable-line react-hooks/exhaustive-deps
-
-  useEffect(() => {
-    return () => {
-      if (stream) {
-        stream.getTracks().forEach((track) => track.stop());
-      }
-    };
-  }, [stream]);
-
-  useEffect(() => {
-    if (selectedAudioDeviceId || selectedVideoDeviceId) {
-      getMedia();
-    }
-  }, [selectedAudioDeviceId, selectedVideoDeviceId]);
-
-  useEffect(() => {
-    if (!socket || !stream) return;
-
-    console.log("Setting up socket event listeners");
-
-    const handleAllUsers = (users: User[]) => {
-      console.log("Received all_users:", users);
-      Object.entries(users).forEach(([socketId, userInfo]) => {
-        console.log("Creating peer connection for:", {
-          socketId,
-          nickname: userInfo.nickname,
-        });
-
-        createPeerConnection(socketId, userInfo.nickname, stream, true, {
-          nickname,
-        });
-      });
-    };
-
-    const handleGetOffer = async (data: {
-      sdp: RTCSessionDescription;
-      offerSendID: string;
-      offerSendNickname: string;
-    }) => {
-      console.log("Received offer from:", data.offerSendID);
-      const pc = createPeerConnection(
-        data.offerSendID,
-        data.offerSendNickname,
-        stream,
-        false,
-        { nickname }
-      );
-      if (!pc) return;
-
-      try {
-        await pc.setRemoteDescription(new RTCSessionDescription(data.sdp));
-        const answer = await pc.createAnswer();
-        await pc.setLocalDescription(answer);
-
-        socket.emit("answer", {
-          answerReceiveID: data.offerSendID,
-          sdp: answer,
-          answerSendID: socket.id,
-        });
-      } catch (error) {
-        console.error("Error handling offer:", error);
-      }
-    };
-
-    const handleGetAnswer = async (data: {
-      sdp: RTCSessionDescription;
-      answerSendID: string;
-    }) => {
-      console.log("Received answer from:", data.answerSendID);
-      const pc = peerConnections.current[data.answerSendID];
-      if (!pc) return;
-      try {
-        await pc.setRemoteDescription(new RTCSessionDescription(data.sdp));
-      } catch (error) {
-        console.error("Error handling answer:", error);
-      }
-    };
-
-    const handleGetCandidate = async (data: {
-      candidate: RTCIceCandidate;
-      candidateSendID: string;
-    }) => {
-      const pc = peerConnections.current[data.candidateSendID];
-      if (!pc) return;
-      try {
-        await pc.addIceCandidate(new RTCIceCandidate(data.candidate));
-      } catch (error) {
-        console.error("Error handling ICE candidate:", error);
-      }
-    };
-
-    const handleReaction = ({
-      senderId,
-      reaction,
-    }: {
-      senderId: string;
-      reaction: string;
-    }) => {
-      if (reactionTimeouts.current[senderId]) {
-        clearTimeout(reactionTimeouts.current[senderId]);
-      }
-
-      if (senderId === socket.id) {
-        setReaction(reaction);
-        reactionTimeouts.current[senderId] = setTimeout(() => {
-          setReaction("");
-          delete reactionTimeouts.current[senderId];
-        }, 3000);
-      } else {
-        addReaction(senderId, reaction);
-        reactionTimeouts.current[senderId] = setTimeout(() => {
-          addReaction(senderId, "");
-          delete reactionTimeouts.current[senderId];
-        }, 3000);
-      }
-    };
-
-    socket.on("all_users", handleAllUsers);
-    socket.on("getOffer", handleGetOffer);
-    socket.on("getAnswer", handleGetAnswer);
-    socket.on("getCandidate", handleGetCandidate);
-    socket.on("user_exit", ({ id }) => closePeerConnection(id));
-    socket.on("room_full", () => {
-      toast.error("해당 세션은 이미 유저가 가득 찼습니다.");
-      navigate("/sessions");
-    });
-    socket.on("reaction", handleReaction);
-
-    return () => {
-      console.log("Cleaning up socket event listeners");
-      socket.off("all_users", handleAllUsers);
-      socket.off("getOffer", handleGetOffer);
-      socket.off("getAnswer", handleGetAnswer);
-      socket.off("getCandidate", handleGetCandidate);
-      socket.off("user_exit");
-      socket.off("room_full");
-      socket.off("reaction", handleReaction);
-
-      if (reactionTimeouts.current) {
-        Object.values(reactionTimeouts.current).forEach((timeout) => {
-          clearTimeout(timeout);
-        });
-      }
-    };
-  }, [
-    socket,
-    stream,
-    nickname,
-    createPeerConnection,
-    closePeerConnection,
-    peerConnections,
-    navigate,
-    toast,
-  ]);
-
-  const emitReaction = (reactionType: string) => {
-    if (socket) {
-      socket.emit("reaction", {
-        roomId: sessionId,
-        reaction: reactionType,
-      });
-    }
-  };
-
-  const joinRoom = async () => {
-    if (!socket || !sessionId || !nickname) {
-      toast.error("닉네임을 입력해주세요.");
-      return;
-    }
-
-    const mediaStream = await getMedia();
-    if (!mediaStream) {
-      toast.error(
-        "미디어 스트림을 가져오지 못했습니다. 미디어 장치를 확인 후 다시 시도해주세요."
-      );
-      navigate("/sessions");
-      return;
-    }
-
-    console.log("Joining room:", sessionId);
-    socket.emit("join_room", { roomId: sessionId, nickname });
-  };
-
-  const addReaction = useCallback(
-    (senderId: string, reactionType: string) => {
-      setPeers((prev) =>
-        prev.map((peer) =>
-          peer.peerId === senderId ? { ...peer, reaction: reactionType } : peer
-        )
-      );
-    }, []); // eslint-disable-line react-hooks/exhaustive-deps
+    joinRoom,
+    emitReaction
+  } = useSession(sessionId);
 
   return (
     <section className="w-screen h-screen flex flex-col max-w-[1440px]">
